fix `PagingResponse` && `module` includes `semestre`

+++ deliberation-service/src/main/java/ma/enset/deliberationservice/dto/session/SessionPagingResponse.java
@@ -9,7 +9,7 @@ public record SessionPagingResponse(
         int page,
         int size,
         int totalPages,
-        int totalElements,
+        long totalElements,
         List<SessionResponse> records
 ) {
 }
