+++ semestre-service/src/main/java/ma/enset/semestreservice/dto/SemestrePagingResponse.java
@@ -6,9 +6,10 @@ import java.util.List;
 
 @Builder
 public record SemestrePagingResponse(
-    int page,
-    int size,
-    int totalPages,
-    int totalElements,
-    List<SemestreResponse> records
-) { }
+        int page,
+        int size,
+        int totalPages,
+        long totalElements,
+        List<SemestreResponse> records
+) {
+}
